msm: vidc: Error handling when core is in bad state
When FW is in bad state, it will not respond to any of the driver's
commands. So all the state transitions will time out. Driver needs
to handle these timeouts properly. When core is in bad state, driver
still needs to free OCMEM. If not, video driver leaks OCMEM clock.

Change-Id: If249c607780ce67d6af2fd5311279152e6281fb7
CRs-Fixed: 725375
Signed-off-by: Praneeth Paladugu <ppaladug@codeaurora.org>
Signed-off-by: Balamurugan Alagarsamy <balaga@codeaurora.org>
Signed-off-by: engstk <eng.stk@sapo.pt>